fix Cache interface definition

+++ lib/PrefixCache.php
@@ -1,14 +1,18 @@
 <?php
+declare(strict_types=1);
 
 namespace Amp\Cache;
 
 use Amp\Promise;
 
-final class PrefixCache implements Cache {
+final class PrefixCache implements Cache
+{
     private $cache;
+
     private $keyPrefix;
 
-    public function __construct(Cache $cache, string $keyPrefix) {
+    public function __construct(Cache $cache, string $keyPrefix)
+    {
         $this->cache = $cache;
         $this->keyPrefix = $keyPrefix;
     }
@@ -18,22 +22,38 @@ final class PrefixCache implements Cache {
      *
      * @return string
      */
-    public function getKeyPrefix(): string {
+    public function getKeyPrefix(): string
+    {
         return $this->keyPrefix;
     }
 
     /** @inheritdoc */
-    public function get(string $key): Promise {
-        return $this->cache->get($this->keyPrefix . $key);
+    public function get($key): Promise
+    {
+        if (!\is_string($key)) {
+            throw new \TypeError("Cache key must be string");
+        }
+
+        return $this->cache->get($this->keyPrefix.$key);
     }
 
     /** @inheritdoc */
-    public function set(string $key, string $value, int $ttl = null): Promise {
-        return $this->cache->set($this->keyPrefix . $key, $value, $ttl);
+    public function set($key, $value, int $ttl = null): Promise
+    {
+        if (!\is_string($key)) {
+            throw new \TypeError("Cache key must be string");
+        }
+
+        return $this->cache->set($this->keyPrefix.$key, $value, $ttl);
     }
 
     /** @inheritdoc */
-    public function delete(string $key): Promise {
-        return $this->cache->delete($this->keyPrefix . $key);
+    public function delete($key): Promise
+    {
+        if (!\is_string($key)) {
+            throw new \TypeError("Cache key must be string");
+        }
+
+        return $this->cache->delete($this->keyPrefix.$key);
     }
 }
